Added data container redesign This is probably very flawed so far - and there are > 200 compile errors hhm,  pcs, datagraph however should work fine Visbricks needs some more work
git-svn-id: svn://localhost:8080/caleydo_main/trunk@4386 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ org.caleydo.testing/src/org/caleydo/testing/collection/VirtualArrayTester.java
@@ -26,7 +26,7 @@ public class VirtualArrayTester extends TestCase {
 	}
 
 	public void testBasics() {
-		assertEquals(size, virtualArray.size().intValue());
+		assertEquals(size, virtualArray.size());
 	}
 
 	public void testBulkRemoveIndex() {
@@ -38,7 +38,7 @@ public class VirtualArrayTester extends TestCase {
 		Collections.shuffle(shuffeledList);
 		virtualArray.removeInBulk(shuffeledList);
 
-		assertEquals(0, virtualArray.size().intValue());
+		assertEquals(0, virtualArray.size());
 	}
 
 //	public void testRemoveID() {
@@ -66,7 +66,7 @@ public class VirtualArrayTester extends TestCase {
 		remove(0, 501, shuffeledList);
 		remove(500, size, shuffeledList);
 
-		assertEquals(0, virtualArray.size().intValue());
+		assertEquals(0, virtualArray.size());
 	}
 
 	private void remove(int from, int to, ArrayList<Integer> shuffeledList) {
